change accrual type fb

+++ internal/core/dto/response/ordersdto/orders.go
@@ -1,8 +1,8 @@
 package ordersdto
 
 type Orders struct {
-	Number     string `json:"number"`
-	Status     string `json:"status"`
-	Accrual    int    `json:"accrual"`
-	UploadedAt string `json:"uploaded_at"`
+	Number     string  `json:"number"`
+	Status     string  `json:"status"`
+	Accrual    float32 `json:"accrual"`
+	UploadedAt string  `json:"uploaded_at"`
 }
